Makefile: Modified the make ubi target to produce large and normal images
Since the kiosk image of LP can be configured to fit into a 256M
flash device and would be made available to customers, in order
to expediate the process of creating both a "normal" and "large"
ubi we decided to edit the build process to produce two final ubi's.
This change also requires the corresponding commit to
device/gateworks/ventana/BoardConfig.mk resulting in two sets of
MKUBI vars.

Signed-off-by: Robert Jones <rjones@gateworks.com>